water sand communication

+++ SandSim/Cell.cpp
@@ -1,8 +1,8 @@
 #include "Cell.h"
 
-Cell::Cell(size_t x, size_t y, CellType type) : x(x), y(y), type(type) {}
+Cell::Cell(int x, int y, CellType type) : x(x), y(y), type(type) {}
 
-bool Cell::isEmpty(size_t nx, size_t ny, size_t width, size_t height) {
+bool Cell::isEmpty(int nx, int ny, int width, int height) {
     return nx >= 0 && nx < width && ny >= 0 && ny < height;
 }
